+++ src/composables/index.ts
@@ -7,5 +7,4 @@ export * from './useTres'
 export * from './useRaycaster'
 export * from './useLogger'
 export * from './useSeek'
-export * from './useRaycaster2'
 export * from './usePointerEventHandler'
