+++ src/composables/useRaycaster/index.ts
@@ -1,59 +1,114 @@
 import { useTres } from '../useTres'
 import { Raycaster, Vector2 } from 'three'
-import { onUnmounted, Ref, ref, ShallowRef, shallowRef } from 'vue'
+import { Ref, computed, onUnmounted, watchEffect } from 'vue'
+import { EventHook, createEventHook, useElementBounding, usePointer } from '@vueuse/core'
 
-/**
- * Raycaster composable return type
- *
- * @export
- * @interface UseRaycasterReturn
- */
-export interface UseRaycasterReturn {
-  /**
-   * Raycaster instance
-   *
-   * @type {ShallowRef<Raycaster>}
-   * @memberof UseRaycasterReturn
-   */
-  raycaster: ShallowRef<Raycaster>
-  /**
-   * Pointer position
-   *
-   * @type {Ref<Vector2>}
-   * @memberof UseRaycasterReturn
-   */
-  pointer: Ref<Vector2>
+export type Intersects = THREE.Intersection<THREE.Object3D<THREE.Event>>[]
+interface PointerMoveEventPayload {
+  intersects?: Intersects
+  event: PointerEvent
 }
 
-/**
- * Composable to provide raycaster support and pointer information
- *
- * @see https://threejs.org/docs/index.html?q=raycas#api/en/core/Raycaster
- * @export
- * @return {*} {UseRaycasterReturn}
- */
-export function useRaycaster(): UseRaycasterReturn {
-  const raycaster = shallowRef(new Raycaster())
-  const pointer = ref(new Vector2())
-  const currentInstance = ref(null)
-  const { setState, state } = useTres()
+interface PointerClickEventPayload {
+  intersects: Intersects
+  event: PointerEvent
+}
 
-  setState('raycaster', raycaster.value)
-  setState('pointer', pointer)
-  setState('currentInstance', currentInstance)
+export const useRaycaster2 = (objects: Ref<THREE.Object3D[]>) => {
+  const { state } = useTres()
 
-  function onPointerMove(event: MouseEvent) {
-    pointer.value.x = (event.clientX / window.innerWidth) * 2 - 1
-    pointer.value.y = -(event.clientY / window.innerHeight) * 2 + 1
+  const canvas = computed(() => state.canvas?.value) // having a seperate computed makes useElementBounding work
+
+  const { x, y } = usePointer({ target: canvas })
+
+  const { width, height, top, left } = useElementBounding(canvas)
+
+  const raycaster = new Raycaster()
+
+  const getRelativePointerPosition = ({ x, y }: { x: number; y: number }) => {
+    if (!canvas.value) return
+
+    return {
+      x: ((x - left.value) / width.value) * 2 - 1,
+      y: -((y - top.value) / height.value) * 2 + 1,
+    }
   }
 
-  state?.renderer?.domElement.addEventListener('pointermove', onPointerMove)
+  const getIntersectsByRelativePointerPosition = ({ x, y }: { x: number; y: number }) => {
+    if (!state.camera) return
+
+    raycaster.setFromCamera(new Vector2(x, y), state.camera)
+
+    return raycaster.intersectObjects(objects.value, false)
+  }
+
+  const getIntersects = (event?: PointerEvent) => {
+    const pointerPosition = getRelativePointerPosition({
+      x: event?.clientX ?? x.value,
+      y: event?.clientY ?? y.value,
+    })
+    if (!pointerPosition) return []
+
+    return getIntersectsByRelativePointerPosition(pointerPosition) || []
+  }
+
+  // const intersects = ref<Intersects>([]) TODO
+
+  // watchEffect(() => {
+  //   intersects.value = getIntersects()
+  // })
+
+  const eventHookClick = createEventHook<PointerClickEventPayload>()
+  const eventHookPointerMove = createEventHook<PointerMoveEventPayload>()
+
+  const triggerEventHook = (eventHook: EventHook, event: PointerEvent) => {
+    eventHook.trigger({ event, intersects: getIntersects(event) })
+  }
+
+  //distinguishing between clicks and drags (in cas of panning or dollying for example) // TODO discuss with other team members
+  let clicked = false
+
+  const onPointerDown = () => {
+    clicked = true
+  }
+  const onPointerMove = (event: PointerEvent) => {
+    clicked = false
+
+    triggerEventHook(eventHookPointerMove, event)
+  }
+
+  const onPointerUp = (event: PointerEvent) => {
+    if (!(event instanceof PointerEvent)) return // prevents triggering twice on mobile devices
+
+    if (clicked) triggerEventHook(eventHookClick, event)
+  }
+
+  const onPointerLeave = (event: PointerEvent) => {
+    eventHookPointerMove.trigger({ event, intersects: [] })
+  }
+
+  const unwatch = watchEffect(() => {
+    if (!canvas?.value) return
+
+    canvas.value.addEventListener('pointerdown', onPointerDown)
+    canvas.value.addEventListener('pointermove', onPointerMove)
+    canvas.value.addEventListener('pointerup', onPointerUp)
+    canvas.value.addEventListener('pointerleave', onPointerLeave)
+
+    unwatch()
+  })
 
   onUnmounted(() => {
-    state?.renderer?.domElement.removeEventListener('pointermove', onPointerMove)
+    if (!canvas?.value) return
+    canvas.value.removeEventListener('pointerdown', onPointerDown)
+    canvas.value.removeEventListener('pointermove', onPointerMove)
+    canvas.value.removeEventListener('pointerup', onPointerUp)
+    canvas.value.removeEventListener('pointerleave', onPointerLeave)
   })
+
   return {
-    raycaster,
-    pointer,
+    // intersects,
+    onClick: (fn: (value: PointerClickEventPayload) => void) => eventHookClick.on(fn).off,
+    onPointerMove: (fn: (value: PointerMoveEventPayload) => void) => eventHookPointerMove.on(fn).off,
   }
 }
