--- src/composables/useRaycaster/useRaycaster.test.ts
@@ -1,23 +1,0 @@
-import { Raycaster, Vector2 } from 'three'
-/* import { useRaycaster } from '.' */
-import { withSetup } from '../../utils/test-utils'
-
-/* const [composable, app] = withSetup(() => useRaycaster()) */
-
-describe.skip('useRaycaster', () => {
-  /*   afterEach(() => {
-    app.unmount()
-  }) */
-  /*   test.skip('provides raycaster', () => {
-    const { raycaster } = composable
-    expect(raycaster).toBeDefined()
-    expect(raycaster.value).toBeInstanceOf(Raycaster)
-  })
-  test.skip('provides pointer', () => {
-    const { pointer } = composable
-    expect(pointer).toBeDefined()
-    expect(pointer.value).toBeInstanceOf(Vector2)
-  }) */
-})
-
-// TODO: find a way to test this with useTresProvider approach
